[1607]: Put the UnvestedCoins function back in view.go and get rid of the locked_coins.go file.
